# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: 群想 AI 群聊 API
  version: 1.0.0
  description: |
    以 backend/src 真实路由为准的接口契约。覆盖认证、短信、群组、消息、文件、
    TTS、社交互动、记忆、Agent、互动分析、监控、画像、人设与用户 API 配置等域。
    部分复杂查询端点（搜索、互动分析、监控指标等）的响应体为宽松对象，
    字段以实现为准，后续逐步收紧。
servers:
  - url: http://localhost:3002
    description: 本地后端
  - url: https://{YOUR_BACKEND_HOST}
    description: 生产后端（占位符，替换为真实 Render/自有域名）
    variables:
      YOUR_BACKEND_HOST:
        default: your-backend-host.example.com
...
security:
  - cookieAuth: []
  - csrfHeader: []
paths:
...
  /api/agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: 获取智能体列表
      responses:
        '200':
          description: 智能体数组
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
    post:
      tags: [Agents]
      operationId: createAgent
      summary: 创建智能体
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreateInput'
      responses:
        '200':
          description: 创建后的智能体
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/agents/generate-questions:
    post:
      tags: [Agents]
      operationId: generateAgentQuestions
      summary: 为智能体配置生成引导问题
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreateInput'
      responses:
        '200':
          description: 问题列表
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/agents/{agentId}:
    get:
      tags: [Agents]
      operationId: getAgent
      summary: 获取智能体详情
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '200':
          description: 智能体详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Agents]
      operationId: updateAgent
      summary: 更新智能体（capabilities 变更会重建模型角色）
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 100 }
                description: { type: string, maxLength: 2000 }
                openingMessage: { type: string, maxLength: 2000 }
                enableSuggestions: { type: boolean }
                capabilities:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: 更新后的智能体
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Agents]
      operationId: deleteAgent
      summary: 删除智能体及其会话
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '200':
          description: 删除结果
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{agentId}/messages:
    get:
      tags: [Agents]
      operationId: listAgentMessages
      summary: 分页获取智能体会话消息
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: before
          schema: { type: string }
      responses:
        '200':
          description: 消息列表
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/agents/{agentId}/chat:
    post:
      tags: [Agents]
      operationId: chatWithAgent
      summary: 与智能体对话
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
              additionalProperties: true
      responses:
        '200':
          description: 智能体回复
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{agentId}/chat-with-files:
    post:
      tags: [Agents]
      operationId: chatWithAgentAndFiles
      summary: 携带文件与智能体对话（multipart）
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              additionalProperties: true
      responses:
        '200':
          description: 智能体回复
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{agentId}/suggestions:
    get:
      tags: [Agents]
      operationId: getAgentSuggestions
      summary: 获取智能体建议问题
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '200':
          description: 建议列表
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
  /api/agents/{agentId}/invoke:
    post:
      tags: [Agents]
      operationId: invokeAgent
      summary: 以结构化输入调用智能体
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 调用结果
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # --------------------------------------------------------------- ApiConfig
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: 登录端点下发的 httpOnly 会话 Cookie
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token
      description: |
        所有非安全方法（POST/PUT/PATCH/DELETE）必须携带，
        令牌通过 GET /api/csrf-token 获取，且须与 Cookie 中的
        CSRF 值一致（double-submit 模式）
  parameters:
...
    AgentIdParam:
      in: path
      name: agentId
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Agent:
      type: object
      required: [id, name, description]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        opening_message: { type: string }
        enable_suggestions: { type: boolean }
        capabilities:
          type: array
          items: { type: string }
        model_roles:
          type: object
          additionalProperties: true
        system_prompt: { type: string }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentCreateInput:
      type: object
      required: [name, description, openingMessage]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 2000
        openingMessage:
          type: string
          minLength: 1
          maxLength: 2000
        enableSuggestions: { type: boolean }
        capabilities:
          type: array
          items: { type: string }
        avatarUrl: { type: string }